- refactor code - calc config should only generate items from given template
